Reorganize searchResult screen
