+++ src/config/locales/en-US.json
@@ -538,7 +538,15 @@
     "more_replies": "more replies"
   },
   "search_result": {
-    "others": "Others",
+    "best": {
+      "title": "Best"
+    },
+    "people": {
+      "title": "People"
+    },
+    "topics": {
+      "title": "Topics"
+    },
     "communities": {
       "title": "Communities",
       "subscribe": "Join",
