+++ src/screens/searchResult/screen/tabs/best/postResultStyles.js
