+++ src/screens/searchResult/screen/tabs/communities/communitiesListItemStyles.js
