+++ src/screens/searchResult/screen/tabs/communities/communitiesListStyles.js
